pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aidos"
version = "0.1.0"
description = "LLM-powered intelligent stress testing agent"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "aidos contributors" }]
keywords = ["stress-testing", "dos", "llm", "agent", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Security",
    "Topic :: Software Development :: Testing :: Traffic Generation",
]
dependencies = [
    "google-genai>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.0.0",
    "click>=8.1.0",
]

[project.scripts]
aidos = "aidos.cli:main"

[project.urls]
Homepage = "https://github.com/aidos-project/aidos"
Repository = "https://github.com/aidos-project/aidos"
Issues = "https://github.com/aidos-project/aidos/issues"
